+++ client/src/components/modals/leo/nameSearchModal.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { useLocation } from "react-router-dom";
 import format from "date-fns/format";
+import { connect } from "react-redux";
 import Modal, { XButton } from "../index";
 import lang from "../../../language.json";
 import State from "../../../interfaces/State";
@@ -8,9 +9,14 @@ import AlertMessage from "../../alert-message";
 import Citizen from "../../../interfaces/Citizen";
 import Weapon from "../../../interfaces/Weapon";
 import Vehicle from "../../../interfaces/Vehicle";
-import { searchName, saveNote, searchNames, suspendLicense } from "../../../lib/actions/officer";
+import {
+  searchName,
+  saveNote,
+  searchNames,
+  suspendLicense,
+  deleteRecordById,
+} from "../../../lib/actions/officer";
 import { Warrant, Ticket, ArrestReport, WrittenWarning } from "../../../interfaces/Record";
-import { connect } from "react-redux";
 import { Item, Span } from "../../../pages/citizen/citizen-info";
 import Select from "../../select";
 
@@ -32,6 +38,7 @@ interface Props {
   saveNote: (citizenId: string, note: string) => void;
   searchNames: () => void;
   suspendLicense: (type: string, citizenId: string) => void;
+  deleteRecordById: (id: string, type: string, citizenId: string) => void;
 }
 
 const NameSearchModal: React.FC<Props> = ({
@@ -41,6 +48,7 @@ const NameSearchModal: React.FC<Props> = ({
   saveNote,
   searchNames,
   suspendLicense,
+  deleteRecordById,
 }) => {
   const [name, setName] = React.useState<any>({});
   const [note, setNote] = React.useState((search && search?.citizen?.note) || "");
@@ -72,6 +80,10 @@ const NameSearchModal: React.FC<Props> = ({
 
   const handleSuspend = (type: string) => () => {
     suspendLicense(type, search?.citizen?.id);
+  };
+
+  const deleteRecord = (id: string, type: string, citizenId: string) => () => {
+    deleteRecordById(id, type, citizenId);
   };
 
   return (
@@ -317,7 +329,7 @@ const NameSearchModal: React.FC<Props> = ({
                             <li
                               key={idx}
                               id={`${idx}`}
-                              className="list-group-item border-dark text-dark"
+                              className="list-group-item border-dark text-dark position-relative"
                             >
                               <Item id="infractions">
                                 <Span>{lang.record.infractions}: </Span>
@@ -343,6 +355,19 @@ const NameSearchModal: React.FC<Props> = ({
                                 <Span>{lang.global.notes}: </Span>
                                 {warning.notes}
                               </Item>
+
+                              <button
+                                type="button"
+                                onClick={deleteRecord(
+                                  warning.id,
+                                  "written_warning",
+                                  warning.citizen_id,
+                                )}
+                                className="btn btn-danger"
+                                style={{ position: "absolute", bottom: "0.2rem", right: "0.2rem" }}
+                              >
+                                {window.lang.global.delete}
+                              </button>
                             </li>
                           );
                         })
@@ -365,7 +390,7 @@ const NameSearchModal: React.FC<Props> = ({
                         search.tickets.map((ticket: Ticket, idx: number) => {
                           return (
                             <li
-                              className="list-group-item border-dark text-dark"
+                              className="list-group-item border-dark text-dark position-relative"
                               key={idx}
                               id={`${idx}`}
                             >
@@ -393,6 +418,15 @@ const NameSearchModal: React.FC<Props> = ({
                                 <Span>{lang.global.notes}: </Span>
                                 {ticket.notes}
                               </Item>
+
+                              <button
+                                type="button"
+                                onClick={deleteRecord(ticket.id, "ticket", ticket.citizen_id)}
+                                className="btn btn-danger"
+                                style={{ position: "absolute", bottom: "0.2rem", right: "0.2rem" }}
+                              >
+                                {window.lang.global.delete}
+                              </button>
                             </li>
                           );
                         })
@@ -415,7 +449,7 @@ const NameSearchModal: React.FC<Props> = ({
                         search.arrestReports.map((report: ArrestReport, idx: number) => {
                           return (
                             <li
-                              className="list-group-item border-dark text-dark"
+                              className="list-group-item border-dark text-dark position-relative"
                               key={idx}
                               id={`${idx}`}
                             >
@@ -443,6 +477,19 @@ const NameSearchModal: React.FC<Props> = ({
                                 <Span>{lang.global.notes}: </Span>
                                 {report.notes}
                               </Item>
+
+                              <button
+                                type="button"
+                                onClick={deleteRecord(
+                                  report.id,
+                                  "arrest_report",
+                                  report.citizen_id,
+                                )}
+                                className="btn btn-danger"
+                                style={{ position: "absolute", bottom: "0.2rem", right: "0.2rem" }}
+                              >
+                                {window.lang.global.delete}
+                              </button>
                             </li>
                           );
                         })
@@ -465,7 +512,7 @@ const NameSearchModal: React.FC<Props> = ({
                         search.warrants.map((warrant: Warrant, idx: number) => {
                           return (
                             <li
-                              className="list-group-item border-dark text-dark"
+                              className="list-group-item border-dark text-dark position-relative"
                               key={idx}
                               id={`${idx}`}
                             >
@@ -625,6 +672,10 @@ const mapToProps = (state: State) => ({
   names: state.officers.names,
 });
 
-export default connect(mapToProps, { searchName, saveNote, searchNames, suspendLicense })(
-  NameSearchModal,
-);
+export default connect(mapToProps, {
+  searchName,
+  saveNote,
+  searchNames,
+  suspendLicense,
+  deleteRecordById,
+})(NameSearchModal);