+++ server/src/routes/records.ts
@@ -149,4 +149,58 @@ router.post(
   },
 );
 
+router.delete(
+  "/:type/:id/:citizenId",
+  useAuth,
+  usePermission(["leo", "dispatch"]),
+  async (req: IRequest, res: Response) => {
+    const { type, id, citizenId } = req.params;
+
+    switch (type) {
+      case "ticket": {
+        await processQuery("DELETE FROM `leo_tickets` WHERE `id` = ?", [id]);
+        break;
+      }
+      case "arrest_report": {
+        await processQuery("DELETE FROM `arrest_reports` WHERE `id` = ?", [id]);
+        break;
+      }
+      case "written_warning": {
+        await processQuery("DELETE FROM `written_warnings` WHERE `id` = ?", [id]);
+        break;
+      }
+      default: {
+        return res
+          .json({
+            error: "invalid type",
+            status: "error",
+          })
+          .status(400);
+      }
+    }
+
+    const citizen = await processQuery("SELECT * FROM `citizens` WHERE `id` = ?", [citizenId]);
+
+    const [vehicles, weapons, warnings, arrestReports, tickets, warrants] = await Promise.all([
+      processQuery("SELECT * FROM `registered_cars` WHERE `citizen_id` = ?", [citizenId]),
+      processQuery("SELECT * FROM `registered_weapons` WHERE `citizen_id` = ?", [citizenId]),
+      processQuery("SELECT * FROM `written_warnings` WHERE `citizen_id` = ?", [citizenId]),
+      processQuery("SELECT * FROM `arrest_reports` WHERE `citizen_id` = ?", [citizenId]),
+      processQuery("SELECT * FROM `leo_tickets` WHERE `citizen_id` = ?", [citizenId]),
+      processQuery("SELECT * FROM `warrants` WHERE `citizen_id` = ?", [citizenId]),
+    ]);
+
+    return res.json({
+      citizen: citizen[0],
+      writtenWarnings: warnings,
+      vehicles,
+      weapons,
+      arrestReports,
+      tickets,
+      warrants,
+      status: "success",
+    });
+  },
+);
+
 export default router;