+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 1.4.9
+
+- New: able to remove tickets/arrest reports
+- Minor: Fixed few bugs
+
 ## 1.4.8
 
 - New: LEO can suspend licenses